% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{isContaminantPrevalence}
\alias{isContaminantPrevalence}
\title{importFrom stats chisq.test
importFrom stats fisher.test}
\usage{
isContaminantPrevalence(freq, neg, method = "auto")
}
\description{
importFrom stats chisq.test
importFrom stats fisher.test
}
\keyword{internal}
