% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abundances.R
\name{plot_abundance}
\alias{plot_abundance}
\title{plot_abundance}
\usage{
plot_abundance(seqtab, taxa_to_plot, conc, normalize = TRUE,
  showModels = TRUE, log = TRUE)
}
\arguments{
\item{seqtab}{(Required). \code{Integer matrix} or \code{phyloseq} object.
A feature table recording the observed abundances of each sequence variant (or OTU) in each sample.
Rows should correspond to samples, and columns to sequences (or OTUs).
If a phyloseq object is provided, the otu-table component will be extracted.}

\item{taxa_to_plot}{(Required). \code{character}.
The names of the taxa to include in this plot.}

\item{conc}{(Required). \code{numeric}. Required if performing frequency-based testing.
A quantitative measure of the concentration of amplified DNA in each sample prior to sequencing.
All values must be greater than zero. Zero is assumed to represent the complete absence of DNA.
If \code{seqtab} was prodivded as a phyloseq object, the name of the appropriate sample-variable in that
phyloseq object can be provided.}

\item{normalize}{(Optional). Default TRUE.
If TRUE, the input \code{seqtab} is normalized so that each row sums to 1 (converted to frequency).
If FALSE, no normalization is performed (the data should already be frequencies or counts from equal-depth samples).}

\item{showModels}{(Optional). Default TRUE.
If TRUE, the contaminant (red) and non-contaminant (black) models are shown in the plot.}

\item{log}{(Optional). Default TRUE.
If TRUE, the axes are log10-scaled.}
}
\description{
plots the abundance of selected taxa in each sample vs. each sample's DNA concentration after PCR amplification
}
\examples{

}
