% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{isContaminant}
\alias{isContaminant}
\title{Identify contaminant sequences.}
\usage{
isContaminant(seqtab, conc, threshold = 0.001, normalize = TRUE,
  detailed = FALSE)
}
\arguments{
\item{seqtab}{(Required). Integer matrix.
A feature table recording the observed abundances of each sequence (or OTU) in each sample.
Rows should correspond to samples, and columns to sequences (or OTUs).}

\item{conc}{(Required). \code{numeric}.
A quantitative measure of the concentration of amplified DNA in each sample prior to sequencing.
All values must be greater than zero. Zero is assumed to represent the complete absence of DNA.}

\item{threshold}{(Optional). Default \code{1e-3}.
The p-value threshold at which the null-hypothesis (not a contaminant) should be rejected in favor of the
alternate hypothesis (contaminant).}

\item{normalize}{(Optional). Default TRUE.
If TRUE, the input \code{seqtab} is normalized so that each row sums to 1 (converted to frequency).
If FALSE, no normalization is performed (the data should already be frequencies or counts from equal-depth samples).}

\item{detailed}{(Optional). Default FALSE.
If TRUE, the return value is a \code{data.frame} containing diagnostic information on the contaminant decision.
If FALSE, the return value is a \code{logical} vector containing the contaminant decisions.}
}
\value{
If \code{detailed=FALSE} a \code{logical} vector is returned, with TRUE indicating contaminants.
If \code{detailed=TRUE} a \code{data.frame} is returned instead.
}
\description{
The frequency of each sequence (or OTU) in the input feature table as a function of the concentration of
amplified DNA in each sample is used to identify contaminant sequences.
}
\examples{
filterAndTrim(testFastqs, filtFastqs, truncQ=2, truncLen=200, rm.phix=TRUE)

}
