% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{isNotContaminant}
\alias{isNotContaminant}
\title{Identify non-contaminant sequences. This is the inverse function of \code{\link{isContaminant}}.}
\usage{
isNotContaminant(seqtab, conc = NULL, neg = NULL, method = "frequency",
  threshold = 0.5, normalize = TRUE, detailed = FALSE)
}
\arguments{
\item{seqtab}{(Required). Integer matrix.
A feature table recording the observed abundances of each sequence (or OTU) in each sample.
Rows should correspond to samples, and columns to sequences (or OTUs).}

\item{conc}{(Optional). \code{numeric}.
A quantitative measure of the concentration of amplified DNA in each sample prior to sequencing.
All values must be greater than zero. Zero is assumed to represent the complete absence of DNA.
REQUIRED if performing frequency-based testing.}

\item{neg}{(Optional). \code{logical}
The negative control samples. Extraction controls give the best results.
REQUIRED if performing prevalence-based testing.}

\item{threshold}{(Optional). Default \code{0.5}.
The p-value threshold at which the null-hypothesis (a contaminant) should be rejected in favor of the
alternate hypothesis (not a contaminant).}

\item{normalize}{(Optional). Default TRUE.
If TRUE, the input \code{seqtab} is normalized so that each row sums to 1 (converted to frequency).
If FALSE, no normalization is performed (the data should already be frequencies or counts from equal-depth samples).}

\item{detailed}{(Optional). Default FALSE.
If TRUE, the return value is a \code{data.frame} containing diagnostic information on the non-contaminant decision.
If FALSE, the return value is a \code{logical} vector containing the non-contaminant decisions.}

\item{method(Optional).}{Default "frequency".
The method used to test for contaminants.
frequency: Contaminants are identified by increased frequency in lower biomass samples.
prevalence: Contaminants are identified by increased prevalence in negative controls.
combined: Both frequency and prevalence are used to identify contaminants.}
}
\value{
If \code{detailed=FALSE} a \code{logical} vector is returned, with TRUE indicating non-contaminants.
If \code{detailed=TRUE} a \code{data.frame} is returned instead.
}
\description{
The frequency of each sequence (or OTU) in the input feature table as a function of the concentration of
amplified DNA in each sample is used to identify non-contaminant sequences. Note that the null hypothesis
here is that sequences **are** contaminants. This function is intended for use on low-biomass samples
in which a large proportion of the sequences are likely to be contaminants.
}
\examples{
isNotContaminant(st, conc, threshold=0.05)

}
